fund_simulation/requirements.txt
streamlit>=1.28.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.17.0
python-dateutil>=2.8.2
openpyxl>=3.1.0
kaleido>=1.1.0
